requirement.txt
python = 3.11.9
ipykernel
jupyter notenook
transformers
torch
librosa
numpy